+++ src/shell/boards/board-context-provider.jsx
@@ -24,7 +24,11 @@ const reducer = (state, action) => {
 			const returnValue = {
 				...state,
 				boards: {
-					[boardKey]: { url: action.payload.url, title: action.payload.title },
+					[boardKey]: {
+						url: action.payload.url,
+						title: action.payload.title,
+						context: action.payload.context
+					},
 					...state.boards
 				},
 				currentBoard: boardKey,
@@ -107,8 +111,11 @@ export default function BoardContextProvider({ children }) {
 	});
 
 	const addBoard = useCallback(
-		(url, title = t('new-tab')) => {
-			dispatch({ type: 'ADD_BOARD', payload: { url, title } });
+		(url, context) => {
+			dispatch({
+				type: 'ADD_BOARD',
+				payload: { url, title: context?.title ?? t('new-tab'), context }
+			});
 		},
 		[t]
 	);